# src/main/resources/openapi/media-api-contract.yaml
openapi: 3.0.3
info:
  title: VaultApp Media API
  description: REST API for managing different types of media (manga, series, films, comics, books, anime)
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
paths:
  /manga:
    get:
      summary: Get all mangas
      description: Retrieve a list of all mangas in the database, with optional pagination.
      operationId: getAllManga
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Page size for pagination
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of mangas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MangaDTO'
        '500':
          description: Internal server error
    post:
      summary: Create a new manga
      description: Adds a new manga to the database.
      operationId: createManga
      requestBody:
        description: Manga object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MangaDTO'
      responses:
        '201':
          description: Manga created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MangaDTO'
        '400':
          description: Invalid manga data
        '500':
          description: Internal server error
  /manga/{id}:
    get:
      summary: Get a specific manga by ID
      description: Retrieve a specific manga by its unique identifier.
      operationId: getMangaById
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the manga to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: The manga object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MangaDTO'
        '404':
          description: Manga not found
        '500':
          description: Internal server error
    put:
      summary: Update a manga
      description: Update the details of an existing manga.
      operationId: updateManga
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the manga to update
          schema:
            type: integer
      requestBody:
        description: Updated manga object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MangaDTO'
      responses:
        '200':
          description: Manga updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MangaDTO'
        '400':
          description: Invalid manga data
        '404':
          description: Manga not found
        '500':
          description: Internal server error
    delete:
      summary: Delete a manga
      description: Deletes a manga by its unique identifier.
      operationId: deleteManga
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the manga to delete
          schema:
            type: integer
      responses:
        '204':
          description: Manga deleted successfully
        '404':
          description: Manga not found
        '500':
          description: Internal server error
  /series:
    get:
      summary: Get all series
      description: Retrieve a list of all series in the database, with optional pagination.
      operationId: getAllSeries
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Page size for pagination
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of series
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SeriesDTO'
        '500':
          description: Internal server error
    post:
      summary: Create a new series
      description: Adds a new series to the database.
      operationId: createSeries
      requestBody:
        description: Series object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeriesDTO'
      responses:
        '201':
          description: Series created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeriesDTO'
        '400':
          description: Invalid series data
        '500':
          description: Internal server error
  /series/{id}:
    get:
      summary: Get a specific series by ID
      description: Retrieve a specific series by its unique identifier.
      operationId: getSeriesById
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the series to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: The series object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeriesDTO'
        '404':
          description: Series not found
        '500':
          description: Internal server error
    put:
      summary: Update a series
      description: Update the details of an existing series.
      operationId: updateSeries
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the series to update
          schema:
            type: integer
      requestBody:
        description: Updated series object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeriesDTO'
      responses:
        '200':
          description: Series updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeriesDTO'
        '400':
          description: Invalid series data
        '404':
          description: Series not found
        '500':
          description: Internal server error
    delete:
      summary: Delete a series
      description: Deletes a series by its unique identifier.
      operationId: deleteSeries
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the series to delete
          schema:
            type: integer
      responses:
        '204':
          description: Series deleted successfully
        '404':
          description: Series not found
        '500':
          description: Internal server error
  /film:
    get:
      summary: Get all films
      description: Retrieve a list of all films in the database, with optional pagination.
      operationId: getAllFilms
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Page size for pagination
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of films
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FilmDTO'
        '500':
          description: Internal server error
    post:
      summary: Create a new film
      description: Adds a new film to the database.
      operationId: createFilm
      requestBody:
        description: Film object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilmDTO'
      responses:
        '201':
          description: Film created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilmDTO'
        '400':
          description: Invalid film data
        '500':
          description: Internal server error
  /film/{id}:
    get:
      summary: Get a specific film by ID
      description: Retrieve a specific film by its unique identifier.
      operationId: getFilmById
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the film to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: The film object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilmDTO'
        '404':
          description: Film not found
        '500':
          description: Internal server error
    put:
      summary: Update a film
      description: Update the details of an existing film.
      operationId: updateFilm
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the film to update
          schema:
            type: integer
      requestBody:
        description: Updated film object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilmDTO'
      responses:
        '200':
          description: Film updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilmDTO'
        '400':
          description: Invalid film data
        '404':
          description: Film not found
        '500':
          description: Internal server error
    delete:
      summary: Delete a film
      description: Deletes a film by its unique identifier.
      operationId: deleteFilm
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the film to delete
          schema:
            type: integer
      responses:
        '204':
          description: Film deleted successfully
        '404':
          description: Film not found
        '500':
          description: Internal server error
...
components:
  schemas:
    MangaDTO:
      $ref: './media-api-schematic.yaml#/components/schemas/MangaDTO'
    SeriesDTO:
      $ref: './media-api-schematic.yaml#/components/schemas/SeriesDTO'
    FilmDTO:
      $ref: './media-api-schematic.yaml#/components/schemas/FilmDTO'
...